Make the maze green coloured too
I did this by making a paletted image that only accepts these 2 shades
of green, and then looping over each pixel of the maze and converting
the high ones (255) to the highest index value from the palette (1).
Doesn't feel very effective but it only has to happen once and I guess
it's fine.