Blink the player on and off
Because you've only got 2 colours and both are already used for walls
and floors, its impossible to tell the player from the maze when the
maze is shown.  Since I can't add more colours, I've gone with animation
and made the player character's pixel blink on/off every half a second.