Hold down movement keys to move continuously
First I tried just letting it move if the key is pressed but the TPS is
so high it moved all across the screen even for a short tap.  I suppose
I could have used it by incrementing a counter and relying on TPS but
instead I made a new ticker that says it's okay to step again after some
milliseconds so that there's time to react but it still moves pretty
fast when pressed down.