Extract movement logic into a Move function
It was becoming very tedious to maintain the long if statements for the
movement logic that were repeating the same thing over and over again
with similar logic.  I've replaced this with a movement function that
does all the checks without relying on X, Y explicitly but instead using
an image.Point to check whether its new location would hit a wall before
allowing it to move forward.  The movement delay and torch toggling has
been moved here too.