Centre smaller mazes in the middle of the screen
To achieve this I wrote a separate constructor for Mazes and a new Maze
object that contains all information about mazes, not just the image but
also the offset, exit and the original maze data which may be useful
later for solving the maze programmatically.  The maze's offset value is
used to centre the maze at Draw time, as well as adjust the position of
the player's pixel accordingly.