Make the step delay after the first keypress longer
To make it easier to make small steps when pressing just once, the first
key press needs a longer delay than the rest to make it easier to tell
between pressing just once and holding it down when it should go fast.

I had some issues with the previous ticker-based solution because the
delay before moving faster would depend on whether you were close in
time to the ticker's tick or not, which is not the indented behaviour.
Instead I've turned the Moving variable into a Step variable that's
continuously counting down to zero.  Its increased when a movement is
initiated and when holding down the key it gets a bigger increase for
the first time round.