Don't move if the key is still being held in from the last level
The new Moved field defaults to false and is reset at each new level.
This tracks whether this is the first movement of the level.  The Move()
method will block movement if this is the first move until the key is
pressed.  This avoids the player from instantly turning off the torch
and running into a surprise wall if they kept the key held as they
finished the last level.  If you keep the key held down it won't do
anything in the new level until you release it and press it again.