Bump the movement counter if the key was pressed in this tick
While the delay is cooling down from the first step it won't bump it up
to "long first step time" if it's still cooling down... so if you tap
multiple times in a row, the first one has a long delay and the next
taps don't.  Now there's a check for if the key was pressed in this
tick, to stop waiting for cooldown, instantly move and bump up the timer
again.

This also provides an opportunity to make the movement delays a bit more
sensitive, so I've adjusted the numbers.