Add new game states with logic for game start
There are now 3 game states:
- Title: the opening screen
- Menu: a game menu (not implemented)
- Level: what has been so far the game itself

The State type is just an integer and the zero value is useful because
it defaults to the title screen which we want to start with.

It also has separate logic for what to draw in the different states and
on transitioning from the title state to the game level state.  This
logic is so small so far that I didn't make new functions for it, but if
they grow I might add an abstractions once it becomes clearer how
they're used.  #YAGNI