Add title screen with graphics from PNG file
This first screen is a proof of concept that I'll expand upon to add
animated graphics in the title screen.  Several shell commands were used
to extract the PNG colour information into an array of 0s and 1s used as
the Pix index array for a paletted title image in the code.  It's quite
convenient because the result is importable Go code.  The next step will
be to write down those commands into a shell script that can run for
many files and then I can use the results for animations.  Initialising
the images might be better done elsewhere than in the main function.