First draft script for converting PNG files to Go source code
The cat command here will need to be replaced with an image magick call
and it still needs some boilerplate around the generated output to make
it valid Go source code but it's a start.  It'll also probably need a
loop of some sorts because the goal is to convert many files, for use as
animation frames.