Finish conversion script
It now generates full Go code with boilerplate based on a provided PNG
filename for input.  I also fixed a bug in the matching, 67 and 199 were
too general, they didn't just match against the numbers but also against
the line numbers, resulting in a black line of zeroes along column 67.