Animate title screen
All Go files are converted from PNG files.  The PNG converter script has
been fixed to generate code that gofmt likes.  A new generate.sh script
calls png2go for each PNG file.  In future it might be replaced with a
Makefile.  It might have saved a lot of time using Statik instead.

There's a new Animation object with the concept of frames and stepping
through them with a delay.  Callers only need to call its Update
function to run this logic and the CurrentFrame function returns an
ebiten Image of the current frame that can be used in your Draw method.