Add animated transition from menu to start of game
This is a lot of animation frames!  Luckily most of the work for
importing frames and abstracting them as an animation has already been
done, so this was fairly easy.  A new state was added to the game logic
TT, for Title Transition.  Now that there's some several state
transitions in the Update function I've noticed I've unintentionally
written a finite state machine, yay.