Fix crash after last level
The next level logic would increase the difficulty level up to the
Extreme level but due to a mistake in the bounds checking (with >= I
included the last level in the level check) the game would crash after
the last level because it would try to increase the difficulty to a
level that doesn't exist.  This was solved by reducing the bounds to
exclude the very last level, so it will keep repeating that one now
instead of crashing.